evaluaciones/requirements.txt
asgiref==3.8.1
Django==5.1
pip-autoremove==0.10.0
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.1
django-ninja==0.17.0
djangorestframework-simplejwt==5.2.2
